pyproject.toml
[project]
name = "mynews-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "google-api-python-client>=2.170.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "litellm>=1.71.1",
    "pytz>=2025.2",
    "requests>=2.32.3",
]